# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for ShortMesh Interface service",
        "title": "Shortmesh - Interface API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "List all devices for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "List all devices",
                "responses": {
                    "200": {
                        "description": "List of devices",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/devices.Device"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new device for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Create a new device",
                "parameters": [
                    {
                        "description": "Device creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devices.CreateDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Requested to add device successfully",
                        "schema": {
                            "$ref": "#/definitions/devices.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Delete a device for the authenticated user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Delete a device",
                "parameters": [
                    {
                        "description": "Device deletion request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devices.DeleteDeviceRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Device deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/devices.DeviceResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/devices/qr-code": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Establishes a WebSocket connection to stream real-time add devices qr-code. This endpoint cannot be tested in Swagger UI - use a WebSocket client instead.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "WebSocket qr-code endpoint (Not executable in Swagger UI)",
                "deprecated": true,
                "responses": {
                    "101": {
                        "description": "WebSocket connection established",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "401": {
                        "description": "Missing or invalid authentication token",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/devices/{device_id}/message": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Queue a message to be sent via the specified device",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "devices"
                ],
                "summary": "Send a message via device",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Device ID",
                        "name": "device_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message to send",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devices.SendMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Message queued successfully",
                        "schema": {
                            "$ref": "#/definitions/devices.SendMessageResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/devices.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "devices.CreateDeviceRequest": {
            "type": "object",
            "required": [
                "platform"
            ],
            "properties": {
                "platform": {
                    "type": "string",
                    "example": "wa"
                }
            }
        },
        "devices.DeleteDeviceRequest": {
            "type": "object",
            "required": [
                "device_id",
                "platform"
            ],
            "properties": {
                "device_id": {
                    "description": "Get the device ID from ListDevices (GET /api/v1/devices) response",
                    "type": "string",
                    "example": "237123456789"
                },
                "platform": {
                    "description": "Get the platform from ListDevices (GET /api/v1/devices) response",
                    "type": "string",
                    "example": "wa"
                }
            }
        },
        "devices.Device": {
            "type": "object",
            "properties": {
                "device_id": {
                    "type": "string",
                    "example": "237123456789"
                },
                "platform": {
                    "type": "string",
                    "example": "wa"
                }
            }
        },
        "devices.DeviceResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Scan the QR code to link your device"
                },
                "qr_code_url": {
                    "type": "string",
                    "example": "/api/v1/devices/qr-code"
                }
            }
        },
        "devices.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "message"
                }
            }
        },
        "devices.SendMessageRequest": {
            "type": "object",
            "required": [
                "contact",
                "platform",
                "text"
            ],
            "properties": {
                "contact": {
                    "type": "string",
                    "example": "1234567890"
                },
                "platform": {
                    "description": "Get the platform from ListDevices (GET /api/v1/devices) response",
                    "type": "string",
                    "example": "wa"
                },
                "text": {
                    "type": "string",
                    "example": "Hello, World!"
                }
            }
        },
        "devices.SendMessageResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Message queued successfully"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter your token in the format: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}